# contracts/openapi/runner-v1.yaml
openapi: 3.1.0
info:
  title: Runner Service API
  version: 1.0.0
  description: |
    The public, versioned contract of the Runner Service.

    The Runner is a standalone bounded service. It **executes** submitted Test
    IR; it does not author test cases and does not generate Test IR from
    natural language. A Test Authoring / IR Generation service integrates with
    the Runner through this document alone — there is no supported source-code
    dependency in either direction, and a caller may be written in any language.

    Three consumption styles are equally supported after submitting a run:
    poll `GET /api/v1/executions/{executionId}`, receive a webhook via
    `callback`, or consume the SSE stream at `/events`.
servers:
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/v1/registry/elements/{elementId}:
    get:
      tags: [Registry]
      summary: Get a registry element
      description: |
        `workspaceRef` is required on every Registry read. An element belonging
        to another workspace answers `404`, never `403`: saying "forbidden"
        would confirm that the id exists somewhere.
      parameters:
        - name: elementId
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/WorkspaceRefQuery'
      responses:
        '200':
          description: The registry element.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistryElementV1'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/registry/elements:
    get:
      tags: [Registry]
      summary: Search registry elements
      description: |
        Searches by display name, alias, system name and description, in that
        order of authority. Each result carries `matchedOn` and `matchScore`, so
        a caller choosing between candidates can tell a human-authored name from
        an inferred description match — which a bare list cannot express.
      parameters:
        - $ref: '#/components/parameters/WorkspaceRefQuery'
        - name: query
          in: query
          description: Free text matched against names, aliases and descriptions.
          schema: { type: string }
        - name: role
          in: query
          description: |
            Narrows to elements declaring this role. An element with no recorded
            role is kept, so an entry that predates the hint is not excluded.
          schema: { type: string }
        - name: pageId
          in: query
          schema: { type: string }
        - name: componentId
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: Matching elements, strongest match first.
          content:
            application/json:
              schema:
                type: object
                required: [elements]
                properties:
                  elements:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/RegistryElementV1'
                        - type: object
                          required: [matchedOn, matchScore]
                          properties:
                            matchedOn:
                              type: string
                              enum: [ID, DISPLAY_NAME, ALIAS, SYSTEM_NAME, DESCRIPTION]
                            matchScore:
                              type: number
                              minimum: 0
                              maximum: 1
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/registry/pages:
    get:
      tags: [Registry]
      summary: List registry pages
      parameters:
        - $ref: '#/components/parameters/WorkspaceRefQuery'
      responses:
        '200':
          description: The workspace's pages.
          content:
            application/json:
              schema:
                type: object
                required: [pages]
                properties:
                  pages:
                    type: array
                    items:
                      type: object
                      required: [pageId, displayName, urlPatterns, revision]
                      properties:
                        pageId: { type: string }
                        displayName: { type: string }
                        description: { type: string }
                        urlPatterns:
                          type: array
                          items: { type: string }
                        revision: { type: integer }
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/registry/components:
    get:
      tags: [Registry]
      summary: List registry components
      parameters:
        - $ref: '#/components/parameters/WorkspaceRefQuery'
        - name: pageId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: The workspace's components.
          content:
            application/json:
              schema:
                type: object
                required: [components]
                properties:
                  components:
                    type: array
                    items:
                      type: object
                      required: [componentId, displayName, revision]
                      properties:
                        componentId: { type: string }
                        displayName: { type: string }
                        description: { type: string }
                        pageId: { type: string }
                        revision: { type: integer }
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/registry/modifications:
    get:
      tags: [Registry]
      summary: List pending registry changes
      description: |
        Drafts and proposals awaiting a decision. Every Registry change is a
        draft first — including one proposed by self-healing — so this is where
        a reviewer sees what is waiting. Decided modifications are terminal and
        do not appear here.
      parameters:
        - $ref: '#/components/parameters/WorkspaceRefQuery'
      responses:
        '200':
          description: Pending modifications.
          content:
            application/json:
              schema:
                type: object
                required: [modifications]
                properties:
                  modifications:
                    type: array
                    items:
                      type: object
                      required: [id, entityKind, type, status, proposedBy, createdAt]
                      properties:
                        id: { type: string }
                        entityKind: { type: string, enum: [ELEMENT, PAGE, COMPONENT] }
                        entityId: { type: string }
                        type:
                          type: string
                          enum:
                            - SELECTOR_UPDATE
                            - RENAME
                            - DESCRIPTION_UPDATE
                            - ALIAS_ADD
                            - ALIAS_REMOVE
                            - ELEMENT_CREATE
                            - ELEMENT_DELETE
                            - AVAILABILITY_UPDATE
                        status: { type: string, enum: [DRAFT, PROPOSED] }
                        proposedBy:
                          type: string
                          enum: [USER, AI, HEALING, RECORDER, SYSTEM]
                        reason: { type: string }
                        createdAt: { type: string, format: date-time }
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/registry/revisions/{entityId}:
    get:
      tags: [Registry]
      summary: Read an entity's change history
      description: |
        Oldest revision first. `version` is monotonic per entity and is recorded
        on each step result, so a past run stays explainable after the Registry
        moves on.
      parameters:
        - name: entityId
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/WorkspaceRefQuery'
      responses:
        '200':
          description: The entity's revisions.
          content:
            application/json:
              schema:
                type: object
                required: [revisions]
                properties:
                  revisions:
                    type: array
                    items:
                      type: object
                      required: [id, version, entityKind, changeType, changedBy, timestamp]
                      properties:
                        id: { type: string }
                        version: { type: integer }
                        entityKind: { type: string }
                        changeType: { type: string }
                        changedBy:
                          type: string
                          enum: [USER, AI, HEALING, RECORDER, SYSTEM]
                        modificationId: { type: string }
                        timestamp: { type: string, format: date-time }
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/registry/resolve:
    get:
      tags: [Registry]
      summary: Resolve a target against the Registry alone
      description: |
        Answers whether a named target would resolve, without opening a browser.
        Useful to an IR Generator that wants to check a target before submitting
        a run. Supply `name` or `description`.
      parameters:
        - $ref: '#/components/parameters/WorkspaceRefQuery'
        - name: name
          in: query
          schema: { type: string }
        - name: description
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string }
      responses:
        '200':
          description: |
            The resolution. `resolved: false` with no element means the Registry
            holds nothing matching — not that the request was wrong.
          content:
            application/json:
              schema:
                type: object
                required: [resolved, alternatives]
                properties:
                  resolved: { type: boolean }
                  element:
                    $ref: '#/components/schemas/RegistryElementV1'
                  matchedOn:
                    type: string
                    enum: [ID, DISPLAY_NAME, ALIAS, SYSTEM_NAME, DESCRIPTION]
                  matchScore: { type: number, minimum: 0, maximum: 1 }
                  alternatives:
                    type: array
                    items:
                      type: object
                      required: [elementId, displayName, matchScore]
                      properties:
                        elementId: { type: string }
                        displayName: { type: string }
                        matchScore: { type: number }
        '400':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
...
    WorkspaceRefQuery:
      name: workspaceRef
      in: query
      required: true
      description: |
        Scopes the read. Required rather than defaulted: a lookup that searched
        every workspace would leak one tenant's element names to another.
      schema: { type: string, minLength: 1 }
  responses:
    Error:
      description: A structured Runner error.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/RunnerError'
  schemas:
    RunnerError:
      type: object
      description: |
        `kind` separates a genuine test failure from a setup or infrastructure
        problem, so a reporting system never blames the application under test
        for a Runner-side issue.
      required: [code, kind, message, retryable]
      properties:
        code:
          type: string
          example: ELEMENT_NOT_FOUND
        kind:
          type: string
          enum:
            - TEST_FAILURE
            - PRECONDITION_FAILURE
            - RESOLUTION_FAILURE
            - INFRASTRUCTURE_FAILURE
            - CONTRACT_FAILURE
        message: { type: string }
        retryable: { type: boolean }
        details: { type: object }
...
    RegistryElementV1:
      type: object
      description: |
        The public view of a registry element.

        Note what is absent: the primary selector, its fallbacks and its
        history. Test IR references an element by identity and meaning, never by
        selector, so publishing the mechanics would invite a caller to embed one
        — exactly what the Registry exists to prevent. `elementId` is immutable,
        so IR referencing it survives every rename.
      required:
        [elementId, systemName, displayName, aliases, userConfirmed, confidence, revision, updatedAt]
      properties:
        elementId: { type: string, example: el_1923 }
        systemName:
          type: string
          description: Normalized identifier used when generating Page Object code.
        displayName: { type: string, example: Create Customer Button }
        description: { type: string }
        role: { type: string }
        semanticType: { type: string }
        aliases:
          type: array
          items: { type: string }
          description: Other names this element answers to.
        userConfirmed:
          type: boolean
          description: A human approved this mapping. The strongest confidence signal.
        confidence:
          type: number
          minimum: 0
          maximum: 1
        revision: { type: integer }
        pageId: { type: string }
        componentId: { type: string }
        updatedAt: { type: string, format: date-time }
...